workaround for web2py bug doesn't work

+++ editor/install_layouter.sh
@@ -7,6 +7,10 @@ rm layout -fR
 hg clone https://code.google.com/p/biographer.layout/ --rev experimental layout
 
 # we need only the layout sources
+if [ ! -e 'layout' ]; then
+	echo 'checkout failed'
+	exit 1
+	fi
 cd layout
 rm article example_networks perl pictures -fR
 rm .hg* -fR
