+++ editor/install_visualizer.sh
@@ -7,6 +7,10 @@ rm biographer.visualization -fR
 
 # download
 hg clone --rev ce0138fb634d https://code.google.com/p/biographer.visualization/
+if [ ! -e 'biographer.visualization' ]; then
+	echo 'checkout failed'
+	exit 1
+	fi
 ln biographer.visualization/target/distribution Visualization -s
 cd biographer.visualization
 rm .hg* -fR
